+++ src/main/java/com/andersmmg/falloutstuff/block/custom/StatueBlock.java
@@ -12,24 +12,15 @@ import net.minecraft.world.BlockView;
 
 public class StatueBlock extends AbstractFloorFacingBlock {
 
+    private static final VoxelShape VOXEL_SHAPE = Block.createCuboidShape(5, 0, 5, 11, 10, 11);
+
     public StatueBlock(AbstractBlock.Settings settings) {
         super(settings);
         this.setDefaultState(this.stateManager.getDefaultState().with(FACING, Direction.NORTH));
     }
 
-    private static final VoxelShape VOXEL_SHAPE = Block.createCuboidShape(5, 0, 5, 11, 10, 11);
-    private static final VoxelShape VOXEL_SHAPE_SOUTH = VoxelUtils.rotateShape(Direction.NORTH, Direction.SOUTH, VOXEL_SHAPE);
-    private static final VoxelShape VOXEL_SHAPE_EAST = VoxelUtils.rotateShape(Direction.NORTH, Direction.EAST, VOXEL_SHAPE);
-    private static final VoxelShape VOXEL_SHAPE_WEST = VoxelUtils.rotateShape(Direction.NORTH, Direction.WEST, VOXEL_SHAPE);
-
     @Override
     public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
-        Direction facing = getDirection(state);
-        return switch (facing) {
-            case SOUTH -> VOXEL_SHAPE_SOUTH;
-            case EAST -> VOXEL_SHAPE_EAST;
-            case WEST -> VOXEL_SHAPE_WEST;
-            default -> VOXEL_SHAPE;
-        };
+        return VoxelUtils.rotateShape(getDirection(state), VOXEL_SHAPE);
     }
 }